Split Trajectory and TrajectoryView.
