+++ hf1/arduino/head_controller.cpp
@@ -1,6 +1,6 @@
 #include "head_controller.h"
 #include "servos.h"
-
+#include <Arduino.h>
 #define kHeadTrajeactoryControllerLoopPeriodSeconds 0.03
 
 HeadWaypoint HeadModulatedTrajectoryView::GetWaypoint(int index) const {
@@ -8,7 +8,7 @@ HeadWaypoint HeadModulatedTrajectoryView::GetWaypoint(int index) const {
   // Transform the modulator with the carrier.
   const auto angle_modulator = modulator().state(index).location() * envelope().state(index).location().amplitude();
   return HeadWaypoint(
-    /*seconds=*/index * carrier().interpolation_config().sampling_period_seconds, 
+    /*seconds=*/carrier().seconds(index), 
     HeadTargetState({
       HeadStateVars(
         carrier().state(index).location().pitch() + angle_modulator.pitch(),
@@ -23,7 +23,8 @@ HeadTrajectoryController::HeadTrajectoryController(const char *name)
 
 void HeadTrajectoryController::Update(TimerSecondsType seconds_since_start, int current_waypoint_index) {
   TimerSecondsType time_fraction = (seconds_since_start - trajectory().seconds(current_waypoint_index)) / (trajectory().seconds(current_waypoint_index + 1) - trajectory().seconds(current_waypoint_index));
-  const State ref_position = trajectory().state(current_waypoint_index) + time_fraction * (trajectory().state(current_waypoint_index + 1) - trajectory().state(current_waypoint_index));
+  const State ref_position = (1 - time_fraction) * trajectory().state(current_waypoint_index) + time_fraction * trajectory().state(current_waypoint_index + 1);
+  Serial.printf("t:%f w:%d %f %f\n", seconds_since_start, current_waypoint_index, DegreesFromRadians(ref_position.location().pitch()), DegreesFromRadians(ref_position.location().roll()));
   SetHeadPitchDegrees(DegreesFromRadians(ref_position.location().pitch()));
   SetHeadRollDegrees(DegreesFromRadians(ref_position.location().roll()));
 }
