+++ hf1/arduino/modulated_trajectory_view.hh
@@ -6,6 +6,10 @@ ModulatedTrajectoryView<TState>::ModulatedTrajectoryView(
   : carrier_(carrier),
     modulator_(modulator),
     envelope_(envelope) {
+  // All signals must be resampled with the same number of points than the carrier.
+  const auto period = carrier_.interpolation_config().sampling_period_seconds;
+  modulator_.EnableInterpolation(InterpolationConfig{ .type = InterpolationType::kLinear, .sampling_period_seconds = period });
+  envelope_.EnableInterpolation(InterpolationConfig{ .type = InterpolationType::kLinear, .sampling_period_seconds = period });
 }
 
 template<typename TState>
